give up on the source location scheme for now
my original idea of using the source location information did not work
well because of the infernal multistep sourcemap problem.

however, i have come around to the idea that it is fine to put the
source code itself for this use-case, at least for the time being. I
will use a beautifier and avoid showing line numbers.

at least going to see how that plays out.